% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boobies}
\alias{boobies}
\title{St. Helena Masked Boobies}
\format{
A data frame with 116355 obs. of  6 variables:
\describe{
  \item{track_id}{Unique identifier code for each bird}
  \item{date_gmt}{Character vector representing date (Greenwich Mean Time)}
  \item{time_gmt}{Character vector representing time (Greenwich Mean Time)}
  \item{longitude}{Longitudinal position of bird}
  \item{latitude}{Latitudinal position of bird}
  \item{lon_colony}{Longitudinal position of breeding colony}
  \item{lat_colony}{Latitudinal position of breeding colony}
  ...
}
}
\source{
\url{https://link.springer.com/article/10.1007/s00265-015-1903-3}
}
\usage{
boobies
}
\description{
A GPS tracking data set of Masked Boobies during incubation and chick-rearing at St. Helena Island. 
Formatted following BirdLife International's Seabird Tracking Database standard \url{www.seabirdtracking.org}.
}
\keyword{datasets}
