% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSpaceUse.r
\name{estSpaceUse}
\alias{estSpaceUse}
\title{Estimate the space use of tracked animals using kernel utilization distribution}
\usage{
estSpaceUse(DataGroup, Scale = 50, UDLev = 50, Res = NULL,
  polyOut = FALSE, plot = FALSE)
}
\arguments{
\item{DataGroup}{data.frame or SpatialPointsDataFrame. Must have an 'ID' field; if input is data.frame or unprojected SpatialPointsDF, must also include 'Latitude' and 'Longitude' fields. For formatting issues see \code{\link{formatFields}}.}

\item{Scale}{numeric (in kilometers). The smoothing parameter ('H') used in the kernel density estimation, which defines the width of the normal distribution around each location. The \code{\link{findScale}} function can assist in finding sensible scales.}

\item{UDLev}{numeric (percent). Specify which utilization distribution contour to show in the plotted output. NOTE: this will only affect the output if \code{polyOut=TRUE}.}

\item{Res}{numeric (in kilometers). Grid cell resolution (in square kilometers) for kernel density estimation. Default is a grid of 500 cells, with spatial extent determined by the latitudinal and longitudinal extent of the data.}

\item{polyOut}{logical scalar (TRUE/FALSE). If TRUE then output will include a plot of individual UD polygons and a simple feature with kernel UD polygons for the level of \code{UDLev}. NOTE: creating polygons of UD is both computationally slow and prone to errors if the usage included in \code{UDLev} extends beyond the specified grid. In this case \code{estSpaceUse} will return only the \code{estUDm}object and issue a warning.}

\item{plot}{logical scalar (TRUE/FALSE). If TRUE, map will be produced showing core areas, each level of ID with a different color. NOTE:\code{polyout} must be TRUE for this to work.}
}
\value{
Returns an object of class \code{estUDm} which is essentially a list, with each item representing the utilization distribution of a level of 'ID'. Values in the output signify the usage probability per unit area for that individual in each grid cell. This can be converted into a SpatialPixelsDataFrame via the \link[adehabitatHR]{estUDm2spixdf} function.

If \code{polyOut=TRUE} the output will be a list with two components: \emph{'KDE.Surface'} is the \code{estUDm} object and \code{UDPolygons} is polygon object of class \code{sf} (Simple Features) with the UD contour for each individual at the specified \code{UDLev}.

If \code{polyOut=TRUE} but the polygon delineation in \code{adehabitatHR::getverticeshr} fails, output is an object of class \code{estUDm} and a warning will be issued.
}
\description{
\code{estSpaceUse} is a wrapper for \code{\link{kernelUD}} which estimates the utilization distribution (UD) of multiple individuals or tracks in a tracking dataset.
}
\details{
A utilization distribution will be calculated for each unique 'ID'. The data should be regularly sampled or interpolated (see adehabitatLT package for functions to this end).
}
\examples{
\dontrun{UDs <- estSpaceUse(DataGroup=Trips, Scale = HVALS$mag, UDLev = 50, polyOut=T)}

}
\seealso{
\code{\link{formatFields}}, \code{\link{tripSplit}}, \code{\link{findScale}}
}
