% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findKBA.r
\name{findKBA}
\alias{findKBA}
\title{Delineating areas of aggregation of tracked animals to identify potential Key Areas for Biodiversity (KBA).}
\usage{
findKBA(KDE.Surface, Represent, Col.size = NA, UDLev = 50,
  plotit = TRUE)
}
\arguments{
\item{KDE.Surface}{estUDm. Either as created by \code{\link{estSpaceUse}} or \code{adehabitatHR::kernelUD}. Only accepted if the utilization distribution was calculated in a projected coordinate reference system.}

\item{Represent}{Numeric (between 0-1). Output value provided by \code{\link{repAssess}} which assesses how representative the tracking data are for characterising the space use of the wider population. If this value is <0.7 then a warning will be issued as the data do not meet the representativeness criteria for a KBA.}

\item{Col.size}{Numeric, the number of individuals breeding or residing at the origin location from where animals were tracked, quantifying the population that the tracking data represent. This number will be used to calculate how many animals use the delineated areas of aggregation. If no value for \code{Col.size} is provided then output will be as the proportion of the population.}

\item{UDLev}{Numeric (percentage). Specifies the quantile used for delineating the core use areas of individuals based on the kernel density distribution. Default set to 50\% based on Lascelles et al. (2016). For penguins higher values can be accepted, see Dias et al. (2018).}

\item{plotit}{Logical. If TRUE then a map of identified areas will be drawn.}
}
\value{
Returns an object of class \code{sf} containing polygon data with three data columns:
  Column \code{N_IND} indicates the number of tracked individuals whose core use area (at \code{UDLev}) overlapped with this polygon.

  Column \code{N_animals} estimates the number of animals from the represented population that regularly use the polygon area. If no value for (at \code{Col.size}) was provided, this number is in \% of the size of the represented population.

  Column \code{KBA} indicates whether the polygon can be considered a potential KBA.

  If \code{plotit=T}, a map is produced, displaying the areas which hold aggregations above a certain threshold proportion of the population. If there are no areas displayed on the map, then either the species doesn't aggregate, the Scale is too small to identify aggregations in this species, or the tracked sample aren't representative enough to meet the thresholds.
}
\description{
\code{findKBA} uses the utilization distributions of individual animals to identify areas of aggregation (i.e. where a large proportion of individuals' core areas overlap).
}
\details{
The function first calculates thresholds based on the representativeness of the tracking data (as quantified by \code{\link{repAssess}}). \code{findKBA} then summarises the number of individual core UDs overlapping in an area and compares that number against the thresholds.
The areas identified are POTENTIAL Key Biodiversity Areas. That is, they are areas of ecological relevance to the species, but must yet be assessed against global criteria (conservation status and global population size of the species) to determine whether they achieve global (or regional) KBA status. All criteria are published in the KBA standard, which may be found here: \url{http://www.keybiodiversityareas.org/what-are-kbas}.
}
\examples{
\dontrun{
findKBA(KDE.Surface, Represent=Represent$out)
}
}
