% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findScale.R
\name{findScale}
\alias{findScale}
\title{Find an appropriate smoothing parameter}
\usage{
findScale(Trips, ARSscale = T, Colony, Res = 100,
  Trips_summary = NULL)
}
\arguments{
\item{Trips}{SpatialPointsDataFrame or data.frame of animal relocations as formatted by \code{\link{formatFields}}. Must include 'ID' field. If input is a data.frame or unprojected SpatialPointsDF, must also include 'Latitude' and 'Longitude' fields.}

\item{ARSscale}{logical scalar (TRUE/FALSE). Do you want to calculate the scale of area-restricted search using First Passage Time analysis? NOTE: does not allow for duplicate date-time stamps.}

\item{Colony}{data.frame with 'Latitude' and 'Longitude' columns specifying the locations of the central place (e.g. breeding colony) from which distances will be calculated.}

\item{Res}{numeric (in kilometers). The desired grid cell resolution (square kilometers) for subsequent kernel analysis (NOT performed in this function). If this is not specified, the scale of movement is compared to a 500-cell grid, with spatial extent determined by the latitudinal and longitudinal extent of the data.}

\item{Trips_summary}{data.frame. Output of \code{\link{tripSummary}} function. If not specified, \code{\link{tripSummary}} will be called within the function.}
}
\value{
Returns a one-row dataframe with smoothing parameter ('H') values and the foraging range estimated from the data set.

This function returns a one-row dataframe with the foraging range in the first column (i.e. 'med_max_distance') calculated by \code{\link{tripSummary}}. The following columns contain various candidate smoothing parameter ('h') values calculated in the following ways:
\enumerate{
  \item 'mag' - log of the foraging range, calculated as the median maximum trip distance
  \item 'scaled_mag' - \eqn{med_max_distance / mag}
  \item 'href' - reference bandwidth a simple, data-driven method which takes into account the number of points, and the variance in X and Y directions.

   \eqn{sqrt((X + Y)* (n^(-1/6)))}; where X=Longitude/Easting, Y=Latitude/Northing, and n=number of relocations
  \item 'scaleARS' - spatial scale of area-restricted Search behavior as estimated using First Passage Time analysis (see \code{\link[adehabitatLT]{fpt}})
}
}
\description{
\code{findScale} takes a tracking data set and outputs a series of candidate smoothing parameter values. Additionaly, it compares the scale of movement resolved by the sampling resolution of the data set, to a grid of desired resolution.
}
\details{
The purpose of this function is to provide guidance regarding the two most sensitive steps in the track2KBA analysis: specification of the (1) smoothing parameter and the (2) grid cell size for kernel density estimation (KDE). Specifically, the goal is to allow for exploration of the effect of these parameters and their inter-relatedness, so that an informed decision may be made regarding their specification in subsequent track2KBA steps.

Kernel density estimation has been identified as particularly sensitive to the specification of the smoothing parameter ((AKA bandwidth, or 'H' value), that is, the parameter that defines the width of the normal distribution around each location. Many techniques for identifying 'optimal' smoothing parameters have been proposed (see Gitzen, Millspaugh, and Kernohan for a classic review; see Fleming and Calabreses 2017 for a later implementation) and many of these techniques have their merits; however, in the track2KBA implementation of KDE we have opted for simplicity.

In the context of the track2KBA analysis, the smoothing parameter ought to represent the relevant scale at which the animal interacts with the environment. Therefore, when selecting a \emph{Scale} value for subsequent analysis, the user must take into account the movement ecology of the study species. For species which use Area-Restricted Search (ARS) behavior when foraging, First Passage Time analysis may be used to identify the scale of interaction with the environment (Fauchald and Tveraa 2003), however not all species use ARS when foraging and therefore different techniques must be used.

What minimum spatial scales are detectable by the data also depends on the sampling resolution. Therefore, when applying First Passage Time analysis, \code{findScale} sets the range of scales at which movements are analyzed based on the distribution of forward, between-point displacements in the data.

The grid cell size also affects the output of kernel density-based space use analyses. Therefore, by specifying the \emph{Res} parameter you can check whether your desired grid cell size is reasonable, given the scale of movement resolved by your data.
}
\examples{
\dontrun{HVALS <- findScale(Trips, ARSscale = T, Colony = Colony, Trips_summary = trip_distances)}

}
