% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndEffectTest.R
\name{IndEffectTest}
\alias{IndEffectTest}
\title{Test site fidelity}
\usage{
IndEffectTest(Trips, tripID, GroupVar, plot = TRUE, method = c("HR",
  "PHR", "VI", "BA", "UDOI", "HD"), conditional = TRUE, UDLev = 50,
  Scale, grid = 500, nboots = 1000)
}
\arguments{
\item{Trips}{SpatialPointsDataFrame or data.frame. If input is data.frame or unprojected SpatialPointsDF, must include 'Latitude' and 'Longitude' fields.}

\item{tripID}{character. Column in \emph{Trips} corresponding to the within group ID (e.g. trip-individual combination)}

\item{GroupVar}{character. Column in \emph{Trips} corresponding to the between group ID (e.g. individual or track)}

\item{plot}{logical scalar (TRUE/FALSE). Do you want to output a boxplot of the result?}

\item{method}{character. Which method of overlap estimation to use? See \code{\link[adehabitatHR]{kerneloverlap}} for descriptions of each method.}

\item{conditional}{logical scalar (T/F). If TRUE, the function sets to 0 the pixels of the grid over which the UD is estimated, outside the home range of the animal estimated at a level of probability equal to percent. Note that this argument has no effect when meth="HR" (from \code{\link[adehabitatHR]{kerneloverlap}}).}

\item{UDLev}{numeric. The desired contour level of the utilization distribution to be used in overlap estimation. NOTE: this is irrelevant if \emph{conditional=FALSE}.}

\item{Scale}{numeric (in kilometers). Smoothing ('H') parameter for kernel density estimation.}

\item{grid}{numeric or SpatialPixels. If numeric, specify the desired number of grid cells over which the utilization distributions will be esimated. A default grid of 500 cells is used.}

\item{nboots}{numeric. Indicate the desired number of Kolmogorov-Smirnov iterations to run. 500 is an advisable minimum for statistical rigor.}
}
\value{
\code{IndEffectTest} returns a list containing three objects. In the first slot 'Overlap Matrix', the full matrix of overlap comparisons. In the 'Overlap' slot, a dataframe with a column identifying whether each overlap estimate corresponds to a within-group, or a between-group comparison. In the third slot 'Kolmogorov-Smirnov' is the test output of the Kolmogorov-Smirnov test, indicating the D parameter and significance estimates.
}
\description{
\code{IndEffectTest} tests whether the variance in overlap between space use areas within a group (e.g within individuals) is significant compared to between groups (e.g. between individuals).
}
\details{
This function works by producing kernel density areas at a desired contour level (i.e. \emph{UDLEv}) for each level of \emph{tripID} and estimating the degree of overlap between all pairwise comparisons using the desired overlap \emph{method}. Then, comparisons are split into 'within' and 'between' groups, determined by the grouping variable (i.e \emph{GroupVar}) argument.

If \emph{conditional=TRUE} then the overlap estimates will range from 0 to \emph{UDLev} (unless \emph{method="HR"}).

Then, the empirical distribution of each group is compared in a bootstrapped Kolmogorov-Smirnov test, to check whether differences in the distributions are significant. If so, it indicates that individuals within the \emph{GroupVar} reuse sites more than expected by chance.

NOTE: Because \code{IndEffectTes} relies on \code{\link[adehabitatHR]{kerneloverlap}} to estimate overlap, it was not possible to implement a \emph{Res} argument as is done in other track2KBA functions. Therefore, it is advised to either leave the default of 500 cells, or ascertain the number of cells in the grid of chosen \emph{Res} from the output of \link{estSpaceUse}.
}
\examples{
\dontrun{
indEffect <- IndEffectTest(Trips, GroupVar="ID", tripID="trip_id", method="BA", Scale=HVALS$mag)

indEffect$`Kolmogorov-Smirnov`}

}
