% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KDE_example}
\alias{KDE_example}
\title{Utilization distributions examples}
\format{
An "estUDm" object:
\describe{
  \item{h}{Smoothing parameter in meters}
  \item{proj4string}{Spatial projection}
  ...
}
}
\usage{
KDE_example
}
\description{
Four Utilization Distributions derived for four Masked Boobies
using track2KBA::estSpaceUse and a scale parameter of 30 km.
}
\keyword{datasets}
