% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapKDE.R
\name{mapKDE}
\alias{mapKDE}
\title{Make simple maps of Kernel Density Estimates}
\usage{
mapKDE(KDE, Colony = NULL, Show = TRUE)
}
\arguments{
\item{KDE}{Simple feature MULTIPOLYGON or estUDm object. Must be output of \code{\link{estSpaceUse}} function).}

\item{Colony}{data.frame. Optional.'Latitude' and 'Longitude' locations to display reference point of, for example, a breeding or tagging site.}
}
\value{
Returns a figure of either single map with all core ranges displayed together, or a series of facetted maps, each of which shows a utilization distribution corresponding to a level of ID in \emph{KDE}.
}
\description{
\code{mapKDE} uses output from \code{estSpaceUse} to create maps illustrating utilization distributions for each ID.
}
\details{
If the input is simple features polygons, these will be displayed for all IDs on same map. If input estUDm utilization distribution surface, each ID level gets its own facet displaying the full UD.
}
\examples{
\dontrun{kde_maps <- mapKDE(Trips)}
}
\seealso{
\code{\link{estSpaceUse}}
}
