% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTracks.R
\name{projectTracks}
\alias{projectTracks}
\title{Project tracking data}
\usage{
projectTracks(dataGroup, projType, custom, reproject = TRUE)
}
\arguments{
\item{dataGroup}{data.frame or SpatialPointsDataFrame. Tracking data, with 
fields as named by \code{\link{formatFields}}.}

\item{projType}{character. Select type of equal-area projection to use. Two 
options are are available: 'cylin' projects to a World Cylindrical Equal Area
projection, and 'azim' projects to a Lambert Azimuthal EA.}

\item{custom}{logigal (TRUE/FALSE). Choose whether projection will use
default centering parameters or whether to set projection center on centroid
of latitude and longitude in dataGroup.}

\item{reproject}{logical (TRUE/FALSE). If your dataGroup dataframe is already
 projected, would you like to reproject these to a custom equal-area 
 projection?

Input data can be tracks split into trips (i.e. output of 
\code{\link{tripSplit}}).}
}
\value{
Returns a SpatialPointsDataFrame, which can be used for the following
 functions: \code{\link{findScale}}, \code{\link{estSpaceUse}}, 
 \code{\link{indEffectTest}}, \code{\link{repAssess}} .
}
\description{
\code{projectTracks} A convenience function to project tracking data to a an 
equal-area projection for use in kernel density analysis.
}
\details{

}
\examples{
\dontrun{

data(boobies)

tracks <- formatFields(boobies, BLformat=TRUE)

tracks_prj <- project(tracks)

}

}
\seealso{
\code{\link{tripSummary}}
}
