% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripSplit.r
\name{tripSplit}
\alias{tripSplit}
\alias{splitSingleID}
\title{Split tracking data into trips}
\usage{
tripSplit(tracks, Colony, InnerBuff = 15, ReturnBuff = 45,
  Duration = 12, Nests = FALSE, plotit = T, rmColLocs = T,
  cleanDF = F)

splitSingleID(Track, Colony, InnerBuff = 15, ReturnBuff = 45,
  Duration = 12, Nests = FALSE, proj.UTM)
}
\arguments{
\item{tracks}{data.frame. Must contain 'Latitude', 'Longitude', 'ID' and 'DateTime' columns (correct format may be assured using \code{\link{formatFields}} function).}

\item{Colony}{data.frame. Containing 'Latitude' and 'Longitude' fields specifying the central location(s) from which trips begin. If data are from MoveBank this information may be extracted using the \code{\link{move2KBA}} function. If \emph{Nests}=T, each row should correspond to an appropriate location (Lat/Lon) for each ID value in \emph{tracks}.}

\item{InnerBuff}{numeric (in kilometers). Indicate the distance that an animal must travel for the movement to be considered a trip. Note that this is also the metric that determines whether two subsequent trips are split - if your animal records locations > \code{InnerBuff} (km) from its place of origin and no locations at the place of origin (e.g. for burrow-nesting species) then subsequent trips may be lumped into a single trip. Increase \code{InnerBuff} to ensure correct splitting of trips.}

\item{ReturnBuff}{numeric (in kilometers). Indicate the proximity required for a trip to be considered as returning. This is useful for identifying incomplete trips (i.e. where data storage/transmission failed during the trip).}

\item{Duration}{numeric (in hours). The period of time that the animals must be at large for the movement to be considered a trip.}

\item{Nests}{logical scalar (TRUE/FALSE). Should the central place used in trip-splitting be specific to each ID? If so, each place must be matched with an 'ID' value in both \emph{tracks} and \emph{Colony} objects.}

\item{plotit}{logical scalar (T/F). Should trips be plotted? If so, the first 20 will be vizualized.}

\item{rmColLocs}{logical scalar (T/F). Should points associated with the central location (e.g. colony) be filtered out of the output?}

\item{cleanDF}{logical scalar (T/F). Should columns which are non-essential for track2KBA analysis be removed from dataframe, or not? Removal will speed analysis up a bit.}

\item{Track}{dataFrame. If using singleSplitID() directly, 'ID' field not needed.}

\item{proj.UTM}{CRS (Coordindate Reference System) object. As generated by sp::CRS()}
}
\value{
Returns a projected, SpatialPointsDataFrame, with the field 'trip_id' added to identify each unique trip-ID combination. If rmColLocs=T, then output has been filtered of points deemed not associated with trip movements.
}
\description{
\code{tripSplit} employs \code{splitSingleID} to split data from multiple individuals' into discrete trips made from centrally-located places.
}
\details{
This function splits central place foraging animal movement data into individual trips away from a central location based on distance and time.

\emph{Nests}=T may be used if it is desired, for example, to use specific nest locations instead of one central location for all individuals/tracks.
}
\examples{
\dontrun{Trips <- tripSplit(tracks, 
Colony=Colony, 
InnerBuff=2, 
ReturnBuff=20, 
Duration=1, 
plotit=T, 
Nests = F, 
rmColLocs = T)}
## needs to be set-up to use published example dataset
}
\seealso{
\code{\link{tripSummary}}
}
