% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFields.R
\name{formatFields}
\alias{formatFields}
\title{Format tracking data for track2BKA analysis}
\usage{
formatFields(tracks, field_ID, field_Lat, field_Lon,
  field_DateTime = NULL, field_Date = NULL, field_Time = NULL,
  format_DT = NULL, cleanDF = FALSE)
}
\arguments{
\item{tracks}{data.frame or data.table.}

\item{field_ID}{character. Unique identifier; e.g. for individuals or tracks.}

\item{field_Lat}{numeric. Name of column corresponding to the LATITUDINAL positions.}

\item{field_Lon}{numeric. Name of column corresponding to the LONGITUDINAL positions.}

\item{field_DateTime}{character. If existing, this is the name of the column corresponding to the combined DATE & TIME.}

\item{field_Date}{character. Name of column corresponding to the DATE only.}

\item{field_Time}{character. Name of column corresponding to the TIME only.}

\item{format_DT}{character. What is the format of the data in your DateTime, Date, and Time columns (e.g. "ymd_HMS")? Specify the format following the standard in \code{\link[lubridate]{parse_date_time}}.}

\item{cleanDF}{logical scalar (T/F). Should columns which are non-essential for track2KBA analysis be removed from dataframe, or not? Removal will speed analysis up a bit.}
}
\value{
Returns a data.frame, with 'ID', 'Latitude', 'Longitude', and 'DateTime' (class POSIXct) columns.
}
\description{
\code{formatFields} formats the column names of a data frame so that they are accepted by track2KBA functions.
}
\details{
By matching up the names of your existing columns with those recognized by track2KBA functions, \code{formatFields} re-formats the data frame, and converts the date/date-time fields into a singe date-time field of class POSIXct.

If date-time is combined in a single column, please use \emph{field_DateTime} instead of \emph{field_Date} and \emph{field_Time}.
}
\examples{
\dontrun{
## using data as formatted on \\url{www.seabirdtracking.org},
i.e. with separate Date and Time fields
 tracks_formatted <- formatFields(tracks, 
 field_ID = "track_id", 
 field_Lat="latitude", 
 field_Lon="longitude", 
 field_Date="date_gmt", 
 field_Time="time_gmt"
 )

## using data with only single Date field
tracks_formatted <- formatFields(tracks, 
field_Lat="lat", 
field_Lon="lon", 
field_Date="Date", 
format_DT = "dmy")
}

}
