% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTrips.R
\name{mapTrips}
\alias{mapTrips}
\title{Make simple maps of foraging trips}
\usage{
mapTrips(Trips, Colony)
}
\arguments{
\item{Trips}{SpatialPointsDataFrame. Must be output of \code{\link{tripSplit}} function).}

\item{Colony}{data.frame. Containing 'Latitude' and 'Longitude' fields specifying the central location(s) from which trips begin. If more than one location, each row should correspond to an appropriate location (Lat/Lon) for each ID value in \emph{Trips}.}
}
\value{
Returns a figure of facetted maps, each of which corresponds to a level of ID in \emph{Trips}.
}
\description{
\code{mapTrips} uses output from \code{tripSplit} to create maps illustrating movements for each ID.
}
\details{
This function only works with the output of \code{tripSplit}.
}
\examples{
\dontrun{Trips <- mapTrips(Trips, Colony)}
}
\seealso{
\code{\link{tripSplit}}
}
