% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTracks.R
\name{projectTracks}
\alias{projectTracks}
\title{Project tracking data}
\usage{
projectTracks(dataGroup, reproject = FALSE)
}
\arguments{
\item{dataGroup}{data.frame or SpatialPointsDataFrame. Tracking data, with fields as named by 
\code{\link{formatFields}}.}

\item{reproject}{logical (TRUE/FALSE). If your dataGroup dataframe is already
 projected, would you like to reproject these to a custom equal-area 
 projection?

Input data can be tracks split into trips (i.e. output of 
\code{\link{tripSplit}}).

Data are transformed to a lambert equal-area projection with it's center 
determined by the data. Data must contain 'Latitude' and 'Longitude' columns.
 Note that this projection may not be the most appropriate for your data and 
 it is almost certainly better to identify a projection appropriate for you 
 study region. So it is not strictly necessary for \code{projectTracks} to be
  used in track2KBA analysis, what is important is that an equal-area 
  projection of some kind is used when constructing kernel density estimates.}
}
\value{
Returns a SpatialPointsDataFrame, which can be used for the following
 functions: \code{\link{findScale}}, \code{\link{estSpaceUse}}, 
 \code{\link{indEffectTest}}, \code{\link{repAssess}} .
}
\description{
\code{projectTracks} Projects tracking data to a custom lambert equal-area 
projection for use in kernel density analysis.
}
\examples{
\dontrun{

data(boobies)

tracks <- formatFields(boobies, BLformat=TRUE)

tracks_prj <- project(tracks)

}

}
\seealso{
\code{\link{tripSummary}}
}
