% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTracks.R
\name{projectTracks}
\alias{projectTracks}
\title{Project tracking data}
\usage{
projectTracks(Tracks, Reproject = FALSE)
}
\arguments{
\item{Tracks}{data.frame. Tracking data, with fields named as named by \code{\link{tripSummary}}.}

\item{Reproject}{logical (TRUE/FALSE). If your Tracks dataframe is already projected, would you like to reproject these to a custom equal-area projection?
Data are transformed to a lambert equal-area projection with it's center determined by the data. Data must contain 'Latitude' and 'Longitude' columns. It is not strictly necessary for this projection to be used in track2KBA analysis, what is important is that an equal-area projection, of any kind, is used when constructing kernel density estimates.}
}
\value{
Returns a SpatialPointsDataFrame, which can be used for the following functions: \code{\link{findScale}}, \code{\link{estSpaceUse}},  \code{\link{IndEffectTest}},  \code{\link{repAssess}} .
}
\description{
\code{projectTracks} Projects tracking data to a custom equal-area projection for use in track2KBA analysis.
}
\examples{
\dontrun{

data(boobies)

tracks <- formatFields(boobies, BLformat=TRUE)

tracks_prj <- project(tracks)

}

}
\seealso{
\code{\link{tripSummary}}
}
