% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repAssess.r
\name{repAssess}
\alias{repAssess}
\title{Assess sample representativeness}
\usage{
repAssess(DataGroup, Scale = 10, Iteration = 50, Res = 100,
  BootTable = T, Ncores = 1)
}
\arguments{
\item{DataGroup}{SpatialPointsDataFrame or data.frame of animal relocations. Must include 'ID' field. If input is data.frame or unprojected SpatialPointsDF, must also include 'Latitude' and 'Longitude' fields.}

\item{Scale}{numeric. This value sets the smoothing (h) parameter for Kernel Density Estimation.}

\item{Iteration}{numeric. Number of times to repeat sub-sampling procedure.}

\item{Res}{numeric. Grid cell resolution (in square kilometers) for kernel density estimation. Default is a grid of 500 cells, with spatial extent determined by the latitudinal and longitudinal extent of the data.}

\item{BootTable}{logical (TRUE/FALSE). Do you want to save the full results table to the working directory?}

\item{Ncores}{numeric. The number of processing cores to use in parallel processing. Check how many are available with parallel::detectCores().}
}
\value{
A single-row data.frame, with columns '\emph{SampleSize}' signifying the maximum sample size in the data set, '\emph{out}' signifying the percent representativeness of the sample, '\emph{type}' is the type of asymptote value used to calculate the '\emph{out}' value, and '\emph{asym}' is the asymptote value used.

There are three potential valules for '\emph{type}': 'asymptote' is the ideal, where the asymptote value is calculated from the parameter estimates of the successful nls model fit. Sometimes nls fit is successful but the curve either flips or does not approach 0.5. In these cases the 'type' is 'asymptote_adj' and the inclusion rate is compared to an artificially adjusted asymptote value of 0.5. Finally, when nls fails, then the mean inclusion rate is taken for the largest sample size; 'type'=='inclusion.'
}
\description{
\code{repAssess} estimates the degree to which the space use of a tracked sample of animals represents that of the larger population.
}
\details{
Representativeness is assessed by calculating the proportion of out-sample points included in in-sample space use areas.

First, the data set is iteratively sub-sampling, producing a pooled utilization distribution (via Kernel Density Estimation), and isolating the 50\% contour. Then, the un-selected points are overlayed on the 50\% contour area and the proportion which overlap is calculated. A representative dataset would approach an inclusion rate of 0.5.
By fitting a trend line to the relationship between sample size inclusion rate we can identify the sample size at which the curve approaches an asymptote, signifying that any new data would simply add to existing knowledge.
}
\examples{
\dontrun{repr <- repAssess(Trips, Scale=10, Iteration=1, BootTable = F, n.cores = 1)}

}
