% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripSplit.r
\name{tripSplit}
\alias{tripSplit}
\alias{splitSingleID}
\title{Split tracking data into trips}
\usage{
tripSplit(tracks, Colony, InnerBuff = NULL, ReturnBuff = NULL,
  Duration = NULL, Nests = FALSE, rmNonTrip = TRUE)

splitSingleID(Track, Colony, InnerBuff = 15, ReturnBuff = 45,
  Duration = 12, Nests = FALSE)
}
\arguments{
\item{tracks}{data.frame. Must contain 'Latitude', 'Longitude', 'ID' and 'DateTime' columns (correct format may be assured using \code{\link{formatFields}} function).}

\item{Colony}{data.frame. Containing 'Latitude' and 'Longitude' fields specifying the central location(s) from which trips begin. If data are from MoveBank this information may be extracted using the \code{\link{move2KBA}} function. If \emph{Nests}=T, each row should correspond to an appropriate location (Lat/Lon) for each ID value in \emph{tracks}.}

\item{InnerBuff}{numeric (in kilometers). Indicate the distance that an animal must travel for the movement to be considered a trip. Note that this is also the metric that determines whether two subsequent trips are split - if your animal records locations > \code{InnerBuff} (km) from its place of origin and no locations at the place of origin (e.g. for burrow-nesting species) then subsequent trips may be lumped into a single trip. Increase \code{InnerBuff} to ensure correct splitting of trips.}

\item{ReturnBuff}{numeric (in kilometers). Indicate the proximity required for a trip to be considered as returning. This is useful for identifying incomplete trips (i.e. where data storage/transmission failed during the trip).}

\item{Duration}{numeric (in hours). The period of time that the animals must be at large for the movement to be considered a trip.}

\item{Nests}{logical scalar (TRUE/FALSE). Should the central place used in trip-splitting be specific to each ID? If so, each place must be matched with an 'ID' value in both \emph{tracks} and \emph{Colony} objects.}

\item{rmNonTrip}{logical scalar (TRUE/FALSE). Should periods not associated with trips be filtered out? Note that this does not filter out the trip starting and ending points which fall within InnerBuff, to allow more accurate calculations of duration and distance covered with \code{tripSummary}. Default is TRUE.}

\item{Track}{dataFrame. If using singleSplitID() directly, 'ID' field not needed.}
}
\value{
Returns an un-projected (WGS84) SpatialPointsDataFrame, with the field 'trip_id' added to identify each unique trip-ID combination. If rmNonTrip=TRUE, then output has been filtered of points deemed not associated with trip movements.
}
\description{
\code{tripSplit} employs \code{splitSingleID} to split data from multiple individuals' into discrete trips made from centrally-located places.
}
\details{
This function splits central place foraging animal movement data into individual trips away from a central location based on distance and time.

\emph{Nests}=T may be used if it is desired, for example, to use specific nest locations instead of one central location for all individuals/tracks.
}
\examples{
\dontrun{Trips <- tripSplit(tracks, 
Colony=Colony, 
InnerBuff=2, 
ReturnBuff=20, 
Duration=1, 
Nests = F, 
rmNonTrip = T)}
}
\seealso{

}
