% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripSummary.r
\name{tripSummary}
\alias{tripSummary}
\title{Summary of trip movements}
\usage{
tripSummary(Trips, Colony = Colony, Nests = FALSE)
}
\arguments{
\item{Trips}{projected SpatialPointsDataFrame. Specifically, as produced by \code{\link{tripSplit}}.}

\item{Colony}{data.frame with 'Latitude' and 'Longitude' columns specifying the locations of the central place (e.g. breeding colony). If Nests=TRUE, Colony should have a third column, 'ID' with corresponding character values in the 'ID' field in \emph{Trips}.}

\item{Nests}{logical scalar (TRUE/FALSE). Were central place locations used in \code{tripSplit} specific to each unique 'ID'? If so, each place must be matched with an 'ID' value in both \emph{Trips} and \emph{Colony} objects.}
}
\value{
Returns a tibble data.frame grouped by ID. Trip characteristics included are trip duration (in hours), distances (in kilometers), directions, start and end times as well as a unique trip identifier ('trip_id') for each trip performed by each individual in the data set. Distances are great circle as calculated by \code{\link[geosphere]{distGeo}}.
}
\description{
\code{tripSummary} provides a simple summary of foraging trips made by central place foraging animals.
}
\details{
\emph{Nests}=T may be used if it is desired, for example, to use specific nest locations instead of one central location for all individuals/tracks.
}
\seealso{
\code{\link{tripSplit}}
}
